Fixed table bugs
Table variables are now manually metatabled to fix a bug that caused them to be passed as reference to all instances of the component that contained them.

Also fixed some visual bugs with displaying tables in the editor hierarchy.

NOTE: Private variables do not exist in Lua, if developers want to hide variables from the editor they MUST currently declare them using the "self.[]" syntax in the "OnStart" function.