Remove Function and Table as key functionality
Removed the ability to use tables and functions as keys in tables and have them display in the editor. This is to simplify the process of reading and writing the scene values to Json.

Note: this functionality still works within Lua, but those values cannot be saved as overwritten parameters in the scene json.